+++ src/ElasticSan/ElasticSan.Autorest/custom/Update-AzElasticSanVolumeGroup.ps1
@@ -107,6 +107,19 @@ function Update-AzElasticSanVolumeGroup {
 
         [Parameter()]
         [Microsoft.Azure.PowerShell.Cmdlets.ElasticSan.Category('Body')]
+        [System.Int32]
+        # The number of days to retain the resources after deletion.
+        ${DeleteRetentionPolicyRetentionPeriodDay},
+
+        [Parameter()]
+        [Microsoft.Azure.PowerShell.Cmdlets.ElasticSan.PSArgumentCompleterAttribute("Enabled", "Disabled")]
+        [Microsoft.Azure.PowerShell.Cmdlets.ElasticSan.Category('Body')]
+        [System.String]
+        # .
+        ${DeleteRetentionPolicyState},
+
+        [Parameter()]
+        [Microsoft.Azure.PowerShell.Cmdlets.ElasticSan.Category('Body')]
         [System.Boolean]
         # A boolean indicating whether or not Data Integrity Check is enabled
         ${EnforceDataIntegrityCheckForIscsi},