+++ src/ElasticSan/ElasticSan/ChangeLog.md
@@ -18,6 +18,16 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for soft delete on volume group and volumes
+    - `New-AzElasticSanVolumeGroup`
+    - `Get-AzElasticSanVolumeGroup`
+    - `Get-AzElasticSanVolume`
+    - `Remove-AzElasticSanVolume`
+    - `Restore-AzElasticSanVolume`
+
+* Added support for whether disk snapshot backup can be taken for volumes.
+    - `Test-AzElasticSanVolumeBackup`
+    - `Test-AzElasticSanVolumeRestore`
 
 ## Version 1.2.1
 * Upgraded nuget package to signed package.